Make sleep a float & add more sorts

+++ config.go
@@ -9,7 +9,7 @@ type Config struct {
 	LIST_LENGTH int `json:"list_length"` //the length of the list
 	BLOCK_WIDTH int `json:"block_width"` //the width of each block
 	BLOCK_HEIGHT_MULT int `json:"block_height_mult"` //the amout each block's height is multiplied by when displayed
-	SLEEP int `json:"sleep"` //the amount of sleep time between showings
+	SLEEP float64 `json:"sleep"` //the amount of sleep time between showings
 	BG [4]uint8 `json:"bg"` //the background color, in RGBA, with each number being between 0 and 0xff
 	FG [4]uint8 `json:"fg"` //the foreground color, as above
 	CHANGED [4]uint8 `json:"changed"` //the color of changed blocks, as above
